Is this even called an commit?
Read the changes...

+++ BFarm.cs
@@ -4,11 +4,11 @@ namespace CookieClicker
 {
     static public class BFarm
     {
-        public static int quantity = 0;
+        public static double quantity = 0;
         public static double initCps = 8;
-        public static int initCost = 1100;
+        public static double initCost = 1100;
         public static double cps = 0;
-        public static int cost = 1100;
+        public static double cost = 1100;
         public static bool[] upgrades = new bool[1000];
         public static void Buy(int qt)
         {
